Chatbot_Foundora/requirements.txt
langchain
langchain-google-genai
langgraph
langchain-community
python-dotenv
streamlit
